Radiology-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
python-multipart==0.0.6
pytorch==2.1.0
openvino==2023.2.0
transformers==4.35.0
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
pydicom==2.4.3
opencv-python==4.8.1.78
tensorflow==2.14.0
scikit-learn==1.3.2 